Refactor ClusterPointIndicesDecomposer with ~sort_by param
