--- jsk_pcl_ros/include/jsk_pcl_ros/cluster_point_indices_decomposer_z_axis.h
@@ -1,58 +1,0 @@
-// -*- mode: C++ -*-
-/*********************************************************************
- * Software License Agreement (BSD License)
- *
- *  Copyright (c) 2013, Ryohei Ueda and JSK Lab
- *  All rights reserved.
- *
- *  Redistribution and use in source and binary forms, with or without
- *  modification, are permitted provided that the following conditions
- *  are met:
- *
- *   * Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- *   * Redistributions in binary form must reproduce the above
- *     copyright notice, this list of conditions and the following
- *     disclaimer in the documentation and/o2r other materials provided
- *     with the distribution.
- *   * Neither the name of the JSK Lab nor the names of its
- *     contributors may be used to endorse or promote products derived
- *     from this software without specific prior written permission.
- *
- *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- *  FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- *  COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- *  BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- *  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- *  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- *  LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- *  POSSIBILITY OF SUCH DAMAGE.
- *********************************************************************/
-
-#ifndef JSK_PCL_ROS_CLUSTER_POINT_INDICES_DECOMPOSER_Z_ORDER_H_
-#define JSK_PCL_ROS_CLUSTER_POINT_INDICES_DECOMPOSER_Z_ORDER_H_
-
-#include <ros/ros.h>
-#include <ros/names.h>
-
-#include "jsk_recognition_msgs/ClusterPointIndices.h"
-#include "sensor_msgs/PointCloud2.h"
-
-#include "jsk_pcl_ros/cluster_point_indices_decomposer.h"
-
-namespace jsk_pcl_ros
-{
-  class ClusterPointIndicesDecomposerZAxis: public ClusterPointIndicesDecomposer
-  {
-  public:
-    virtual void sortIndicesOrder(pcl::PointCloud<pcl::PointXYZ>::Ptr input,
-                                  std::vector<pcl::IndicesPtr> indices_array,
-                                  std::vector<pcl::IndicesPtr> &output_array);
-  };
-}
-
-#endif
