+++ jsk_pcl_ros/sample/include/play_rosbag_arc2017_4objects.xml
@@ -1,0 +1,33 @@
+<launch>
+
+  <param name="/use_sim_time" value="true" />
+
+  <node name="rosbag_play"
+        pkg="rosbag" type="play"
+        args="$(find jsk_pcl_ros)/sample/data/arc2017_4objects.bag --clock --loop">
+  </node>
+
+  <group ns="/camera/rgb">
+    <node name="republish"
+          pkg="image_transport" type="republish"
+          args="compressed raw">
+      <remap from="in" to="image_raw" />
+      <remap from="out" to="image_raw" />
+    </node>
+  </group>
+  <group ns="/camera/depth_registered">
+    <node name="republish"
+          pkg="image_transport" type="republish"
+          args="compressedDepth raw">
+      <remap from="in" to="image_raw" />
+      <remap from="out" to="image_raw" />
+    </node>
+  </group>
+  <include file="$(find openni2_launch)/launch/openni2.launch">
+    <arg name="camera" value="camera"/>
+    <arg name="publish_tf" value="false" />
+    <arg name="load_driver" value="false"/>
+    <arg name="depth_registration" value="true"/>
+  </include>
+
+</launch>
